Move modal content and form content to general components folder, fix sidebar active button in subroutes
