+++ src/components/Modal/ModalContent/NewPlan/Footer/index.tsx
